Refactor accordion to accept a fold strategy
The accordion and squeeze functions now accept an `AccordionFoldStrategy` to calculate the number of folds. A default strategy is added, as well as two additional strategies: `Fixed` to have a fixed number of folds, and `FoldSize` to determine the fold count based on a specified size.
The `n` parameter was removed from accordion.
Sample app was updated to use the new parameters.